gpu kernel int

+++ add.cu
@@ -43,7 +43,7 @@ int main() {
   cudaEventRecord(start);
 
   for (int i = 0; i < niter; i++) {
-    gpu_kernel(iter, [] GPU_LAMBDA (float a, float b) {
+    gpu_kernel(iter, [] GPU_LAMBDA (int a, float b) {
       return a + b;
     });
   }
